pyproject.toml
[project]
name = "ESTREAMER"
version = "0.1.0"
description = "Web server for analyising and processing live stream chat"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-decouple (>=3.8,<4.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "flask (>=3.1.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "google-cloud-pubsub (>=2.33.0,<3.0.0)",
    "beartype (>=0.22.9,<0.23.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "yt-dlp (>=2025.12.8,<2026.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"
ruff = "^0.14.9"
